Add functions.py and changes in tests
